+++ src/app/_services/authentication.service.ts
@@ -7,6 +7,7 @@ import {Role} from '../_models/role';
 import {User} from '../_models/user';
 import {CollectionService} from './collection.service';
 import {DashboardService} from './dashboard.service';
+import {CopyService} from './copy.service';
 
 @Injectable({providedIn: 'root'})
 export class AuthenticationService {
@@ -14,8 +15,7 @@ export class AuthenticationService {
     public currentUser: Observable<User>;
 
     constructor(private http: HttpClient,
-                protected collectionService: CollectionService,
-                protected dashboardService: DashboardService) {
+                protected copyService: CopyService) {
         this.currentUserSubject = new BehaviorSubject<User>(JSON.parse(localStorage.getItem('currentUser')));
         this.currentUser = this.currentUserSubject.asObservable();
     }
@@ -41,8 +41,7 @@ export class AuthenticationService {
     logout() {
         // remove user from local storage to log user out
         localStorage.removeItem('currentUser');
-        this.collectionService.clearCopied();
-        this.dashboardService.clearCopied();
+        this.copyService.clearCopied();
         this.setCurrentUser(null);
     }