+++ src/app/_services/copy.service.spec.ts
@@ -1,0 +1,19 @@
+import {TestBed} from '@angular/core/testing';
+
+import {CopyService} from './copy.service';
+import {RouterTestingModule} from '@angular/router/testing';
+import {HttpClientTestingModule} from '@angular/common/http/testing';
+
+describe('CopyService', () => {
+    beforeEach(() => {
+        TestBed.configureTestingModule({
+            imports: [RouterTestingModule, HttpClientTestingModule],
+            providers: [CopyService]
+        });
+    });
+
+    it('should be created', () => {
+        const service: CopyService = TestBed.get(CopyService);
+        expect(service).toBeTruthy();
+    });
+});